requirements.txt
pip
autopep8
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
Werkzeug==3.0.3
requests==2.32.3
google-cloud-storage==2.18.2